+++ src/main/java/net/enzo/plus/common/Config.java
@@ -6,6 +6,7 @@ import java.io.File;
 
 public class Config {
     public static boolean gottaGoFast = true;
+    public static boolean craftOnly = false;
 
     public static void letsConfigurate(File file) {
         Configuration conf = new Configuration(file);
@@ -14,8 +15,9 @@ public class Config {
             conf.load();
 
             gottaGoFast = conf.get("General", "Sonic Speed", gottaGoFast, "Set if Infinity armor gave you sonic speed").getBoolean(true);
+            craftOnly = conf.get("General", "Crafting Only", craftOnly, "Set if the game just accept craftings, for MineTweaker porpouses").getBoolean(false);
         } catch (Exception e) {
-            System.out.println("Plus can't load configs, try reinstall the mod?");
+            System.out.println("The reality is in collapse...");
             e.printStackTrace();
         } finally {
             conf.save();