+++ src/main/java/net/enzo/plus/common/compat/minetweaker/Tweaker.java
@@ -1,0 +1,9 @@
+package net.enzo.plus.common.compat.minetweaker;
+
+import minetweaker.MineTweakerAPI;
+
+public class Tweaker {
+    public static void worstIdeaInMyLife() {
+        MineTweakerAPI.registerClass(AbsoluteCrafting.class);
+    }
+}